requirements.txt
# Core ML and Data Science
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data Collection
requests>=2.25.0
python-decouple>=3.6

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0

# Development and Testing
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0